add news service to api
This change adds the ability to manage the news stories that appear in
the news banner on the new CMAP homepage.

The primary additions are newly defined routes and controllers,
however there are also significant tests and library functions.

Note that a few simple news controllers follow the old paradigm of
try/catch async operations, while other news controllers (notably
updateRanks and createStory) use a new paradigm based on Futures, with
help from the Fluture and Sanctuary libraries.

This change in direction is made out of a desire for stronger type
constraints, and the ability to chain multiple operations together
within those type constraints.

In pursuit of this, this change includes a suite of helper
functions (and tests) that allow a query to be defined in terms of its
template and required parameters, including a type-safe means of
resovling the necessary anguments from the request (or other)
object. In addition a generic controller generator is inculded for
simple operations. Consult the added tests for examples and some
discussion of the shape of the problem and the selected solution.

Other changes include:

create middleware for ensuring correct content type

update log service to set a log threshold, allowing the suppression of
certain startup logs except in production

update package.json with name, description, author; update node engine

update eslintrc: make global variables explicit