use stricter table list for candidate calculation
In order to allow queries with joins and table aliases, do not use the
table references list produced by the query parser's AST to derive a
liste of candidate servers (aliases will not match anything in the
dataset locations map), instead use the result of string parsing which
depends on the CMAP table name conventions.

Additionally, do not return errors to the user if table names do not
match real tables. Return an error if NO references exist in a query,
otherwise let it run.

This changeset also adds tests for two library functions central to
this behavior and fixes incorrect code identifying omitted tables in
compareTableAndDatasetLists.